- OpenCode's plugin loader (v1.15.x) iterates over EVERY export of the plugin so we're splitting this into two files - change the name of the PLATFORM_SOURCE to "opencode" so it can be isolated and its work is not mislabeled as 'claude'
